add repository tests

+++ app/src/main/java/com/dema/store/common/domain/repositories/ProductsRepository.kt
@@ -11,7 +11,7 @@ interface ProductsRepository {
 
     fun getProducts(): Flow<List<Product>> // 1
     fun getCategories(): Flow<List<Category>> // 1
-    suspend fun requestMoreProducts(pageToLoad: Int, numberOfItems: Int,category: Int = 0): PaginatedProducts
+    suspend fun requestMoreProducts(pageToLoad: Int, numberOfItems: Int,category: Long = 0): PaginatedProducts
     suspend fun requestCategories(): List<Category>
     suspend fun storeCategory(category: List<Category>) // 3
     suspend fun storeProducts(category: Category, products: List<ProductWithDetails>) // 3
